deps/v8/test/inspector/debugger/async-stack-created-frame-expected.txt

Checks created frame for async call chain

Running test: testPromise
foo1 (test.js:10:2)
-- Promise.then --
promise (test.js:12:14)
(anonymous) (expr.js:0:0)


Running test: testPromiseThen
foo1 (test.js:10:2)
-- Promise.then --
promiseThen (test.js:20:14)
(anonymous) (expr.js:0:0)

foo2 (test.js:14:2)
-- Promise.then --
promiseThen (test.js:21:14)
(anonymous) (expr.js:0:0)


Running test: testPromiseThenThen
foo1 (test.js:10:2)
-- Promise.then --
promiseThenThen (test.js:29:14)
(anonymous) (expr.js:0:0)

foo1 (test.js:10:2)
-- Promise.then --
promiseThenThen (test.js:30:14)
(anonymous) (expr.js:0:0)

foo2 (test.js:14:2)
-- Promise.then --
promiseThenThen (test.js:29:25)
(anonymous) (expr.js:0:0)


Running test: testPromiseResolve
foo1 (test.js:10:2)
-- Promise.then --
promiseResolve (test.js:36:27)
(anonymous) (expr.js:0:0)


Running test: testPromiseReject
foo1 (test.js:10:2)
-- Promise.catch --
promiseReject (test.js:40:31)
(anonymous) (expr.js:0:0)


Running test: testPromiseAll
foo1 (test.js:10:2)
-- Promise.then --
promiseAll (test.js:44:44)
(anonymous) (expr.js:0:0)


Running test: testPromiseRace
foo1 (test.js:10:2)
-- Promise.then --
promiseRace (test.js:48:45)
(anonymous) (expr.js:0:0)


Running test: testThenableJob1
foo1 (test.js:10:2)
-- Promise.then --
thenableJob1 (test.js:52:72)
(anonymous) (expr.js:0:0)


Running test: testThenableJob2
foo1 (test.js:10:2)
-- Promise.then --
thenableJob2 (test.js:56:57)
(anonymous) (expr.js:0:0)


Running test: testSetTimeouts
foo1 (test.js:10:2)
(anonymous) (test.js:72:25)
-- setTimeout --
(anonymous) (test.js:64:6)
-- setTimeout --
(anonymous) (test.js:63:4)
-- setTimeout --
setTimeouts (test.js:62:2)
(anonymous) (expr.js:0:0)
